Improve TypeRegistry
This changes the behavior of adding types to a TypeRegistry: A common use case for the TypeRegistry is serializing google.protobuf.Any. That typical involves adding the desired types, and the types they use. We can trivially automate the second part of this, making the TypeRegistry more convenient to use.

This also makes the method add() public, and replaces the static methods fromIterable() and fromTypes() with a single from() that takes all supported types as an input, not just message types.