Delete temporary tests
Tried a few options to add them without duplicating the code being tested, but there are no simple options. The function is not exported from the package, so we can't use a separate test package. And to run tests within the package, we need to pull in a bunch of dependencies to transpile for jest.