Clean up descriptor discrimination
This extracts a change from #231 that is independently useful:

`DescriptorSet` provides convenient types to work with descriptors. It uses a row of interfaces like `DescMessage`, `DescEnum`, which are discriminated with a `kind` property.

We made a bad call with `DescField`, which needs further discrimination between map fields, message fields, etc.  We added `kind: "map_field"`, `kind: "message_field"`, etc., which makes it cumbersome to identify fields.

This fixes the issue by using `kind: "field"` for `DescField`, and discriminating further with a second discriminator `fieldKind`.