Improve TS docs on support (#319)
A user encountered [an
issue](https://github.com/bufbuild/protobuf-es/issues/314) with building
because they had `strictNullChecks` set to `false`. We run our
compatibility settings against the defaults generated by TypeScript's
`tsc --init`. The `strict: true` setting is a default, so
`strictNullChecks: false` is actually not an option we should claim to
support.

However, this should be made more clear in our docs and the README has
been updated to reflect that.