Switch code size benchmark to use a JS brotli implementation
We have moved from brotli via zlib to gzip `fflate`, a pure JS implementation in https://github.com/bufbuild/protobuf-es/pull/411.

It allowed us to run diffchecks in CI on the results, but it forced us to use gzip instead of brotli. This PR switches back to brotli, but with a JS implementation instead.