Remove usage of globalThis in generated code (#504)
This removes the usage of `globalThis` in generated code by instead
using an export of global `Number` constants.

This is done so that we can safely use these global constants when
generating code and be assured we're using the correct values. We cannot
use `globalThis` since we support ES2017 and `globalThis` was introduced
in ES2020. We also don't want to explicitly generate code using, for
example, `Number.NaN`, since this could clash with a message name of
`Number`.

Instead we can export them in a local `proto-double.js` file since this
will be in a different scope as the generated code and we are guaranteed
to use the intended global values.

Note that this complements
https://github.com/bufbuild/protobuf-es/pull/488 to remove usage of
`globalThis` in the repo.