Partially improve TypeScript compatibility tests
This partially gets us to the goals defined in #581. We do not remove the dependency on `@types/node` yet, but we make sure to use the corresponding version for every version of TypeScript we test with.

To speed up GitHub actions, the compatibility tests are moved to a separate job that runs in parallel.