Deprecate properties toString and text of JSDocBlock
Currently, the method `GeneratedFile.jsDoc` returns a `JSDocBlock` that leaks implementation details through the properties "toString" and "text".

This PR marks the properties as deprecated, so that we can remove them in the future. This will allow us to accept JSDoc blocks as object literals in `GeneratedFile.print`, in line with similar objects.