Build ESM artifacts with Node16 moduleResolution (#754)
Fixes #599 

This switches all packages to build ESM artifacts with Node16 module
resolution. In doing so, it turns on Node16 for both `moduleResolution`
and `module` by default throughout the repo. It also turns on
`verbatimModuleSyntax` by default (which is why the need for all the
`import type {...}` changes).

For CJS artifacts, the exceptions for `moduleResolution node10` and
`verbatimModuleSyntax false` are now specified in the command script.