Support unknown fields in clone (#758)
Support unknown fields in the method `Message.clone()`. 

Extensions are stored as unknown fields on a message, and cloning a
message will currently discard any set extension. This behavior is
unexpected, since it would be logical that a deep copy includes all
properties that were explicitly set, whether it is a regular field, or
an extension.

This PR fixes this flaw, and includes unknown fields when cloning a
message.

This also documents equal's behavior of ignoring the unknown fields and
extensions.