Add /*@__PURE__*/ annotation for better tree-shaking support
This annotation informs bundlers that the succeeding function call is free of side effects. This means the symbol can be removed from the module during tree-shaking if it is unused.

See https://github.com/bufbuild/protobuf-es/pull/470

The annotation is not terribly effective yet, since all generated descriptors depend on the file descriptor, but it's still worthwhile.