Update reflect
This adds two new types to the reflect API: ReflectList and ReflectMap. The API now always returns messages as ReflectMessage, list fields as ReflectList, and map fields as ReflectMap. Note that accessing an unset message fields returns a ReflectMessage with a new zero-message.

The methods ReflectMessage.addListItem and ReflectMessage.setMapEntry are kept for now. We can decide whether it's worth to keep them for the hot path when we take a look at performance.

This also updates create() to fully support the "init" argument.