Unskip skipped tests
Most importantly, tests for serializing map fields.

Fixes a bug in fromBinary, where map keys were converted to strings, which are the correct representation in a message, but not for a ReflectMap. This bug was most likely introduced in 7054ed75, which limited the conversion to a smaller set of acceptable types.