Remove own FeatureSetDefaults
Remove the options for FeatureSetDefaults from createDescFileSet(), createDescriptorSet(), and @bufbuild/protoplugin.

Going forward, feature resolution will happen internally in createDescFileSet(), and not be exposed. This will avoid several bootstrapping issues and simplify the code overall. Feature resolution with own features can still be implemented as an external function if needed, but at this point, this does not seem to be a crucial feature.