Avoid reserved object property names in generated enumerations
A protobuf enum can have values named "constructor", or "toString". When we generate a TypeScript / JavaScript enumeration for, we append an "$" to the name. These property names are valid in TypeScript, but they can potentially cause issues , depending on how the generated objects are used.

The logic to escape reserved names has a flaw: It is not applied in case the enumeration uses a shared prefix (values each start with the enum name).

This PR fixes the issue and adds test coverage.