Add coverage for serialization with edition2023
This adds coverage for binary and JSON serialization via proto2 and proto3: Since proto2 or proto3 behavior can be configured in editions, we define equivalent messages, and use proto2/proto3 serialization as the source of truth to test serialization with editions.

Note that this does not cover cases that are possible in editions, but not in proto2 or proto3.