V2: Exclude WKT from generated test files
We started generating all well-known types in the package protobuf-test because we wanted to cover both JS and TS generated code.

This is no longer useful with the new runtime. We keep generating google/protobuf/type.proto. It imports other WKT, which lets us test whether pre-generated types are imported as expected.