Automatically publish with the proper dist-tag
npm uses dist-tags to distinguish between the "latest" version, and other distributions. When publishing a package, the tag can be given with the flag `--tag`. When installing a package, the version tagged "latest" will be installed by default. When a tag is specified (for example `npm install foo@alpha`), the tagged version is installed.

Since we follow semver, we can trivially deduce the proper dist-tag from our workspace version: When running the target `release`, we call `npm publish --tag latest` for stable versions such as 3.1.4. For alpha versions such as 2.0.0-alpha.8, we call `npm publish --tag alpha`. Similar for beta versions and release candidates.