Add initPartialPlain to init plain messages with partials
It is possible to run initPartial with a plain message by passing the type
separately from the plain message.

This is useful for initializing plain messages with partials.

In the interest of backwards compatibility. the initPartial function signature
is kept the same. A separate function called initPartialPlain is added which
accepts the type as a third parameter. initPartial is updated to call
initPartialPlain with the type from getType() if source is set.

Signed-off-by: Christian Stewart <christian@aperture.us>