Fix internal function getFieldPresence
Previously, it was hardwired to return LEGACY_REQUIRED for the proto2 `required` label, and EXPLICIT for singular message fields.

The correct behavior for singular message fields is to resolve the feature first, and - only if the feature resolves to IMPLICIT - return EXPLICIT. As a result Edition singular message fields will report LEGACY_REQUIRED if set.