Attach Valid type to codegenv2 descriptor
This introduces codegenv2, with a version of GenMessage that accepts a Valid type, in addition to a JSON type. In protoc-gen-es, we're now generating type expressions using the new gen-type.

We're adding codegenv2 while maintaining codegenv1 so that old gen-code is still compatible with the next version of the runtime.

The types MessageShape, MessageJsonType, etc. are updated to work with both gen-types. We're adding the new export MessageValidType.